python-backend/requirements.txt
amqp==5.1.0
APScheduler==3.9.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
Flask==2.0.3
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.0
Jinja2==3.0.3
kombu==5.2.4
Markdown==3.3.6
MarkupSafe==2.1.0
mysqlclient==2.1.0
packaging==21.3
prompt-toolkit==3.0.28
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.2.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
